Nature armour will no longer heal vampires.

+++ src/main/java/com/minegusta/GearUp/powerlisteners/ShadowBoost.java
@@ -17,7 +17,7 @@ public class ShadowBoost
 
     public boolean isDark()
     {
-        return p.getLocation().getBlock().getLightLevel() < ((byte) 7);
+        return p.getLocation().getBlock().getLightLevel() < 7;
     }
 
     public boolean isInMap()
